Implement unit tests for lsbstego.py Fix bugs identified by unit tests Ensure mask.extractLSB returns the right data type

+++ mask.py
@@ -26,4 +26,4 @@ def applyLSB(data, mask):
 
 
 def extractLSB(data):
-    return data & 1
+    return (data & 1).astype(numpy.uint8)
